refactor(svg): replace unneccessary svg with png
